+++ tests/unit/Transformation/AbstractBaseTransformerTest.php
@@ -3,14 +3,14 @@
 use paslandau\DataFiltering\Transformation\AbstractBaseTransformer;
 use paslandau\DataFiltering\Transformation\DataTransformerInterface;
 
-class BaseTransformerTest extends PHPUnit_Framework_TestCase
+class AbstractBaseTransformerTest extends PHPUnit_Framework_TestCase
 {
 
     public function testNullTrue()
     {
         $expected = null;
-        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class,[null, true]);
-
+        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class, [null, true]);
+        /** @var AbstractBaseTransformer $t */
         $actual = $t->Transform($expected);
         $msg = "Expected: $actual == $expected";
         $this->assertEquals($expected, $actual, $msg);
@@ -20,7 +20,8 @@ class BaseTransformerTest extends PHPUnit_Framework_TestCase
     {
         $this->setExpectedException(get_class(new UnexpectedValueException()));
         $expected = null;
-        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class,[null, false]);
+        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class, [null, false]);
+        /** @var AbstractBaseTransformer $t */
         $t->Transform($expected);
     }
 
@@ -31,16 +32,18 @@ class BaseTransformerTest extends PHPUnit_Framework_TestCase
             return strtoupper($s);
         };
         $mock->expects($this->any())->method("Transform")->will($this->returnCallback($callback));
-        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class,[$mock]);
+        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class, [$mock]);
 
         $input = "hallo";
         $expected = "HALLO";
+        /** @var AbstractBaseTransformer $t */
         $actual = $t->Transform($input);
         $msg = "Expected: $actual == $expected";
         $this->assertEquals($expected, $actual, $msg);
     }
 
-    public function testEnabledCache(){
+    public function testEnabledCache()
+    {
         $calls = 5;
 
         $mock = $this->getMock(DataTransformerInterface::class);
@@ -48,21 +51,24 @@ class BaseTransformerTest extends PHPUnit_Framework_TestCase
             return strtoupper($s);
         };
         $mock->expects($this->once())->method("transform")->will($this->returnCallback($callback));
-        /** @var AbstractBaseTransformer $t */
-        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class,[$mock],'', TRUE, TRUE, TRUE, array('processData'));
+        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class, [$mock], '', TRUE, TRUE, TRUE, array('processData'));
         $t->expects($this->once())->method("processData")->will($this->returnArgument(0));
+
+        /** @var AbstractBaseTransformer $t */
         $t->setIsCacheActive(true);
 
         $input = "hallo";
         $expected = "HALLO";
-        for($i = 0; $i < $calls; $i++){
+        $actual = null;
+        for ($i = 0; $i < $calls; $i++) {
             $actual = $t->transform($input);
         }
         $msg = "Expected: $actual == $expected";
         $this->assertEquals($expected, $actual, $msg);
     }
 
-    public function testDisabledCache(){
+    public function testDisabledCache()
+    {
         $calls = 5;
 
         $mock = $this->getMock(DataTransformerInterface::class);
@@ -70,14 +76,16 @@ class BaseTransformerTest extends PHPUnit_Framework_TestCase
             return strtoupper($s);
         };
         $mock->expects($this->exactly(5))->method("transform")->will($this->returnCallback($callback));
-        /** @var AbstractBaseTransformer $t */
-        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class,[$mock],'', TRUE, TRUE, TRUE, array('processData'));
+        $t = $this->getMockForAbstractClass(AbstractBaseTransformer::class, [$mock], '', TRUE, TRUE, TRUE, array('processData'));
         $t->expects($this->exactly(5))->method("processData")->will($this->returnArgument(0));
+
+        /** @var AbstractBaseTransformer $t */
         $t->setIsCacheActive(false);
 
         $input = "hallo";
         $expected = "HALLO";
-        for($i = 0; $i < $calls; $i++){
+        $actual = null;
+        for ($i = 0; $i < $calls; $i++) {
             $actual = $t->transform($input);
         }
         $msg = "Expected: $actual == $expected";