+++ changelog.md
@@ -1,8 +1,11 @@
-## 0.1.1
+###0.1.1
 
  - updated repositories to local satis installation
+ - added tests for each Transformer
+ - moved parsing logic from HtmlFormTransformer to HtmlForm
+ - implemented fixed parseStr() method to allow "." and " " in form fiels keys
 
-## 0.1.0
+##0.1.0
 
  - changed package name from DataFiltering to data-filtering